requirements.txt
discord.py
python-dotenv
PyNaCl==1.4.0
